Relax requirement in gamma monotonic test
The test was generating a pair of reals `(x, x')`, with `x <= x'`
with `abs (x - x')` so small that `gamma x > gamma x'`, which is
obviously not true.

For now, the solution is to create slightly larger intervals.
The failing test was for the `Double` type, which does not have
boundless precision guarantees.